requirements.txt
kafka-python
faker
pyspark==3.5.0
boto3
pyarrow